+++ src/main/java/com/jerry/pilipala/domain/interactive/handler/CommentInteractiveActionHandler.java
@@ -1,4 +1,4 @@
-package com.jerry.pilipala.domain.vod.service.impl.handler;
+package com.jerry.pilipala.domain.interactive.handler;
 
 import com.jerry.pilipala.domain.vod.entity.mongo.interactive.VodInteractiveAction;
 import com.jerry.pilipala.infrastructure.enums.video.VodInteractiveActionEnum;
@@ -9,25 +9,25 @@ import org.springframework.stereotype.Component;
 import java.util.Map;
 
 @Component
-public class UnfollowInteractiveActionHandler extends InteractiveActionHandler {
-    public UnfollowInteractiveActionHandler(MongoTemplate mongoTemplate,
-                                            RedisTemplate<String, Object> redisTemplate) {
+public class CommentInteractiveActionHandler extends InteractiveActionHandler {
+    public CommentInteractiveActionHandler(MongoTemplate mongoTemplate,
+                                           RedisTemplate<String, Object> redisTemplate) {
         super(mongoTemplate, redisTemplate);
     }
 
     @Override
     public VodInteractiveAction trigger(Map<String, Object> params) {
         VodInteractiveAction interactiveAction = super.trigger(params);
-
         String cid = params.get("cid").toString();
+
         checkVodStatisticsExists(cid);
-        incVodStatistics(cid, "followCount", false);
+        incVodStatistics(cid, "commentCount", true);
 
         return interactiveAction;
     }
 
     @Override
     public VodInteractiveActionEnum action() {
-        return VodInteractiveActionEnum.UNFOLLOW;
+        return VodInteractiveActionEnum.COMMENT;
     }
 }
