+++ src/main/java/com/jerry/pilipala/domain/interactive/handler/LikeInteractiveActionHandler.java
@@ -1,4 +1,4 @@
-package com.jerry.pilipala.domain.vod.service.impl.handler;
+package com.jerry.pilipala.domain.interactive.handler;
 
 import com.jerry.pilipala.domain.vod.entity.mongo.interactive.VodInteractiveAction;
 import com.jerry.pilipala.infrastructure.enums.redis.VodCacheKeyEnum;
@@ -10,20 +10,22 @@ import org.springframework.stereotype.Component;
 import java.util.Map;
 
 @Component
-public class CollectInteractiveActionHandler extends InteractiveActionHandler {
-    public CollectInteractiveActionHandler(MongoTemplate mongoTemplate,
-                                           RedisTemplate<String, Object> redisTemplate) {
+public class LikeInteractiveActionHandler extends InteractiveActionHandler {
+    public LikeInteractiveActionHandler(MongoTemplate mongoTemplate,
+                                        RedisTemplate<String, Object> redisTemplate) {
         super(mongoTemplate, redisTemplate);
     }
 
     @Override
     public VodInteractiveAction trigger(Map<String, Object> params) {
+        // 事件记录
         VodInteractiveAction interactiveAction = super.trigger(params);
 
         String cid = params.get("cid").toString();
-        // 是取消收藏事件，点赞数 - 1，set 移除 uid
+
+        // 是取消点赞事件，点赞数 - 1，set 移除 uid
         String uid = interactiveAction.getUid();
-        String likeSetKey = VodCacheKeyEnum.SetKey.COLLECT_SET.concat(String.valueOf(uid));
+        String likeSetKey = VodCacheKeyEnum.SetKey.LIKE_SET.concat(String.valueOf(uid));
         if (Boolean.TRUE.equals(redisTemplate.opsForSet()
                 .isMember(likeSetKey, cid))) {
             redisTemplate.opsForSet().remove(
@@ -31,24 +33,23 @@ public class CollectInteractiveActionHandler extends InteractiveActionHandler {
                     cid
             );
             checkVodStatisticsExists(cid);
-            incVodStatistics(cid, "collectCount", false);
+            incVodStatistics(cid, "likeCount", false);
         }
-        // 如果是收藏事件，收藏数 +1，uid 放入 set
+        // 如果是点赞事件，点赞数 +1，uid 放入 set
         else {
             redisTemplate.opsForSet().add(
                     likeSetKey,
                     cid
             );
             checkVodStatisticsExists(cid);
-            incVodStatistics(cid, "collectCount", true);
+            incVodStatistics(cid, "likeCount", true);
         }
 
         return interactiveAction;
-
     }
 
     @Override
     public VodInteractiveActionEnum action() {
-        return VodInteractiveActionEnum.COLLECT;
+        return VodInteractiveActionEnum.LIKE;
     }
 }
