+++ src/main/java/com/jerry/pilipala/interfaces/web/DanmakuController.java
@@ -8,7 +8,7 @@ import com.jerry.pilipala.application.vo.vod.DanmakuResponse;
 import com.jerry.pilipala.application.vo.vod.DanmakuValueVO;
 import com.jerry.pilipala.domain.vod.service.DanmakuService;
 import com.jerry.pilipala.domain.vod.service.DanmakuSseManager;
-import com.jerry.pilipala.domain.vod.service.impl.handler.InteractiveActionStrategy;
+import com.jerry.pilipala.domain.interactive.handler.InteractiveActionStrategy;
 import com.jerry.pilipala.infrastructure.enums.video.VodInteractiveActionEnum;
 import io.swagger.annotations.ApiOperation;
 import org.apache.commons.lang3.StringUtils;
